chore: create donwload card button and adjust responsive card
