+++ react-/src/pages/DefaultPage/Components/Card/index.jsx
@@ -2,9 +2,6 @@ import { useContext } from 'react'
 import './styles.css'
 import { SIZES } from './Card.data'
 import { DefaultContextPage } from '../../context'
-import DownloadButton from './DownloadButton'
-import html2canvas from 'html2canvas'
-import downloadjs from 'downloadjs'
 
 const Card = () => {
   const { name, profission, description } = useContext(DefaultContextPage)
@@ -23,13 +20,6 @@ const Card = () => {
     return SIZES.MEDIUM.CLASS
   }
 
-  const downloadCard = async () => {
-    const element = document.querySelector('.section-card')
-    const canvas = await html2canvas(element, { scale: 2 })
-    const dataURL = canvas.toDataURL('image/jpg')
-    downloadjs(dataURL, 'card.jpg', 'image/jpg')
-  }
-
   return (
     <>
       <section className="section-card">
@@ -39,7 +29,6 @@ const Card = () => {
           <p>{description}</p>
         </div>
       </section>
-      <DownloadButton onClick={downloadCard} />
     </>
   )
 }
