+++ react-/src/pages/DefaultPage/Components/Form/DownloadButton/styles.css
@@ -7,7 +7,7 @@
   padding: 5px;
   font-family: 'Kumbh Sans', sans-serif;
   font-weight: bold;
-  width: 180px;
+  width: 100%;
   outline: none;
   text-transform: uppercase;
   transition: 0.5s;
